requirements.txt
flask
opentelemetry-api
opentelemetry-sdk
opentelemetry-ext-http-requests
opentelemetry-ext-wsgi
